+++ datalake/src/main/java/com/sequenceiq/datalake/cm/ClouderaManagerRangerUtil.java
@@ -1,5 +1,6 @@
 package com.sequenceiq.datalake.cm;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
@@ -179,10 +180,10 @@ public class ClouderaManagerRangerUtil {
         }
     }
 
-    public ApiCommand getApiCommand(String stackCrn, int commandId) throws ApiException {
+    public ApiCommand getApiCommand(String stackCrn, long commandId) throws ApiException {
         ApiClient client = clouderaManagerProxiedClientFactory.getProxiedClouderaManagerClient(stackCrn);
         CommandsResourceApi commandsResourceApi = clouderaManagerApiFactory.getCommandsResourceApi(client);
-        return commandsResourceApi.readCommand(commandId);
+        return commandsResourceApi.readCommand(BigDecimal.valueOf(commandId));
     }
 
 }
